requirements.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
execnet==2.0.2
fastapi==0.104.1
fastapi-versioning==0.10.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.27.0
pytest-xdist==3.5.0
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
rsa==4.9
sentry-sdk==1.36.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.16.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.23.2
WTForms==3.1.1
